# modules/apps/domain-api/internal/infrastructure/http/openapi.yaml
openapi: 3.0.3
info:
  title: domain-api
  description: >
    Event-driven CRUD API. Reads are synchronous (straight from
    Postgres). Writes (POST/PUT/DELETE) are accepted asynchronously —
    they publish a command onto the event bus and return 202; the
    change is applied by domain-worker and won't be visible on a GET
    until it's processed.
  version: "1.0.0"
servers:
  - url: /
security:
  - apiKeyAuth: []
paths:
...
  /transacoes:
    get:
      summary: List transacoes by filter
      parameters:
        - name: usuario
          in: query
          required: true
          schema:
            type: string
        - name: conta
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: categoria
          in: query
          required: false
          schema:
            type: string
        - name: de
          in: query
          required: false
          description: Inclusive lower bound, YYYY-MM-DD.
          schema:
            type: string
            format: date
        - name: ate
          in: query
          required: false
          description: Inclusive upper bound, YYYY-MM-DD.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  transacoes:
                    type: array
                    items:
                      $ref: "#/components/schemas/Transacao"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
    post:
      summary: Create a transacao (async — returns 202, not the created row)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransacaoCreateInput"
      responses:
        "202":
          $ref: "#/components/responses/Accepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /transacoes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      summary: Update a transacao (async — returns 202)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransacaoUpdateInput"
      responses:
        "202":
          $ref: "#/components/responses/Accepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
    delete:
      summary: >
        Delete a transacao (async — returns 202). Takes a JSON body,
        same reasoning as DELETE /rooms/{id}.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransacaoDeleteInput"
      responses:
        "202":
          $ref: "#/components/responses/Accepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        Required on every /users route. A valid key also exempts the
        request from the per-IP rate limit that otherwise applies to
        unauthenticated (401) attempts.
  schemas:
...
    Transacao:
      type: object
      properties:
        id:
          type: string
          format: uuid
        usuario_email:
          type: string
        conta_id:
          type: string
          format: uuid
        tipo:
          type: string
          enum: [entrada, saida]
        categoria:
          type: string
        descricao:
          type: string
        anexo_imagem:
          type: string
        valor:
          type: number
          format: double
        data:
          type: string
          format: date
        criado_em:
          type: string
          format: date-time
        atualizado_em:
          type: string
          format: date-time
    TransacaoCreateInput:
      type: object
      required: [usuario_email, conta_id, tipo, categoria, valor, data]
      properties:
        usuario_email:
          type: string
        conta_id:
          type: string
          format: uuid
        tipo:
          type: string
          enum: [entrada, saida]
        categoria:
          type: string
        descricao:
          type: string
        anexo_imagem:
          type: string
        valor:
          type: number
          format: double
          description: Must be greater than zero.
        data:
          type: string
          format: date
    TransacaoUpdateInput:
      type: object
      required: [usuario_email]
      properties:
        usuario_email:
          type: string
        conta_id:
          type: string
          format: uuid
        tipo:
          type: string
          enum: [entrada, saida]
        categoria:
          type: string
        descricao:
          type: string
        anexo_imagem:
          type: string
        valor:
          type: number
          format: double
        data:
          type: string
          format: date
    TransacaoDeleteInput:
      type: object
      required: [usuario_email]
      properties:
        usuario_email:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
    Accepted:
      description: Command accepted — applied asynchronously
      content:
        application/json:
          schema:
            type: object
            properties:
              command_id:
                type: string
                format: uuid
              status:
                type: string
                example: accepted
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: Too many unauthenticated requests from this IP
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"